requirements.txt
pymupdf
pillow
pytesseract
reportlab
streamlit
google-generativeai
